// sources/packages/web/src/types/contracts/InstitutionRouteMeta.ts
export enum InstitutionRouteUserType {
  ADMIN = "Institution_Admin",
  LOCATION_MANAGER = "Institution-location-manager",
  USER = "Institution-user",
}

export enum InstitutionUserTypes {
  // values are same as that of usertype in db
  admin = "admin",
  locationManager = "location-manager",
  user = "user",
}